Improve MLIC and PSD Importer

+++ TexTransCore/BlendTexture/ShaderAsset/AlphaCopy.shader
@@ -41,7 +41,7 @@ Shader "Hidden/AlphaCopy"
                 return o;
             }
 
-            fixed4 frag (v2f i) : SV_Target
+            float4 frag (v2f i) : SV_Target
             {
                 float4 MainColor = tex2Dlod(_MainTex ,float4(i.uv,0,0));
                 float4 AlphaColor = tex2Dlod(_AlphaTex ,float4(i.uv,0,0));
